Initial code commit
This initial commit is under the testbed/ directory which is really a place to play around and test things in before moving over to the real code. Having separate branches might be the best for this going forward but right now its very early in dev anyway.

Currently builds a very basic app using tundra as build system and doesn't rely on CMake and the Qt build system (I don't want to be bound to Qt)

Also requires that Qt5 is installed and QT5 environment is set to the root directory (only Mac OS X tested yet)