Moved ui/imgui -> external/imgui
This is because it's being handled as an external library for now. Might move this back if we do some heavy modifications to it but now with the support of "extra" header in implementation file it makes it a bit easier to extend which should hopefully be quite ok even if we have some locals changes in it.