Smarter sending of register data
When we do a "n" step VICE will send back the registers for us. Thus it's not needed to actually request the registers again so now this is tracked so if we do a step we skip any getRegisters request as it has been completed already.

Also added support for steping and unit-test for it that make sure the test program stayes within the correct location