Added support for serviceFuncs
This allows us to pass functions down into create instance that the plugin can use to request services from the host.
Currently messageFuncs are implement which allows the plugin to show modal message boxes (useful in case of errors, missing files, etc) This can later be extended to ever more functions allowing plugins to be as platform independent as possible.