Return ScImEditor interface from ImScEditText
This change allows external code to send command down to ScEditor. This is needed to in order to control the Scintilla editor as it's based on sending commands to it. This is somewhat annoying as it breaks the statelessness of the ImGui but it's hard to workaround and now the app that is using the code needs to track it instead.

Ref T65