Script cleanup
Various restructring of the script code. Regular functions was inside the the API for scripts which will not work. The whole API is based on struct with function pointers so plugins doesn't need to link with any libs.

This needs to be implemented in another way with a interface provided to the plugins.

Detailed list of changes:

* Moved script/pd_script.cpp -> core/script.cpp
* Moved api/include/pd_script.h -> core/script.h
* Removed script project as it was just one file, can move out from core if it begins to grow
* Removed script test (script tests can be in core tests for now)
* Removed deps on the script project