Moved a bunch of code from bgfx -> imgui_sys
Now some of the imgui wrapped functions are calling Rust code instead. This allows us to handle popup menus directly from Rust instead of going over to C/C++ which would require masharing of data and what not.

Added new imgui_sys create for the Imgui wrapping code. Move can be moved here and the idea is to have Bgfx only dealing with bgfx stuff and imgui_sys dealing with the wrapping and Imgui + some high-level code (such as menu handling)

Also moved private_data pointer for pd_ui to first in the structure. That makes it possible to add new functions at the end of the struct without breaking the ABI/API